requirements.txt
dlib==19.23.0
git+https://github.com/openai/CLIP.git
numpy==1.22.3
opencv-python-headless==4.5.5.64
Pillow==9.1.0
scipy==1.8.0
torch==1.11.0
torchvision==0.12.0
